Add files via upload
the data was obtained from kaggle datasets which is explained as follows: job_title: The job title or role associated with the reported salary. experience_level: The level of experience of the individual. employment_type: Indicates whether the employment is full-time, part-time, etc. work_models: Describes different working models (remote, on-site, hybrid). work_year: The specific year in which the salary information was recorded. employee_residence: The residence location of the employee. salary: The reported salary in the original currency. salary_currency: The currency in which the salary is denominated. salary_in_usd: The converted salary in US dollars. company_location: The geographic location of the employing organization. company_size: The size of the company, categorized by the number of employees.